+++ tests/worker/test_thread.py
@@ -47,28 +47,42 @@ def test_basic_noargs():
     for r,x,y in zip(res, in_x, in_y):
         assert( (x*y) == r )
 
-class TestClass1:
+class Class1:
 
-    def __init__(self):
-        pass
+    def __init__(self, z):
+        self.z = z
+
+    @lox.thread(2)
+    def test_method1(self, x, y):
+        sleep(SLEEP_TIME)
+        return x*y + self.z
 
     @lox.thread
-    def test_method(self, x, y):
+    def test_method2(self, x, y):
         sleep(SLEEP_TIME)
-        return x*y
+        return x*y + self.z
 
-def test_method():
+def test_method_1():
     in_x = [1,   2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12,]
     in_y = [13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24,]
+    z = 5
 
-    test_obj = TestClass1()
+    test_obj = Class1( z )
 
-    assert( 10 == test_obj.test_method(2,5) )
+    assert( (2*5+z) == test_obj.test_method1(2,5) )
+    assert( (2*5+z) == test_obj.test_method2(2,5) )
 
     for x,y in zip(in_x, in_y):
-        test_obj.test_method.scatter(x,y)
-    res = test_obj.test_method.gather()
+        test_obj.test_method1.scatter(x,y)
+    res = test_obj.test_method1.gather()
 
     for r,x,y in zip(res, in_x, in_y):
-        assert( (x*y) == r )
+        assert( (x*y+z) == r )
+
+    for x,y in zip(in_x, in_y):
+        test_obj.test_method2.scatter(x,y)
+    res = test_obj.test_method2.gather()
+
+    for r,x,y in zip(res, in_x, in_y):
+        assert( (x*y+z) == r )